requirements.txt
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Flask==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
MarkupSafe==3.0.2
nltk==3.9.1
python-dotenv==1.1.1
regex==2024.11.6
requests==2.32.4
textblob==0.19.0
tqdm==4.67.1
urllib3==2.5.0
Werkzeug==3.1.3
